services/ingestor/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
PyJWT
cryptography